refactor(layout): move models/callbacks/util into domain/ (final step)
Step 4 of the domain/ reorganization plan -- the final move.
Relocates the last three LCM-internal modules into the domain/
subpackage:

* models.py -> domain/models.py
  Data types: SlotCredential, SlotCode, SyncState, runtime_data,
  the LockCodeManagerConfigEntry type alias.

* callbacks.py -> domain/callbacks.py
  EntityCallbackRegistry. Resolves the directional weirdness the
  auditor flagged: models.py used to import callbacks at the root
  level (a value type pulling on entity-lifecycle infrastructure);
  now they're siblings in domain/ and the import is local.

* util.py -> domain/util.py
  mask_pin, get_slot_coordinator, async_disable_slot.

After this PR the integration root contains only HA-mandated
modules:

* HA platform discovery shims (binary_sensor, event, sensor,
  switch, text) -- can't move per HA's loader contract
* HA-layer entity bases (entity.py)
* Integration plumbing (__init__.py, config_flow.py,
  diagnostics.py, repairs.py, websocket.py)
* Constants (const.py)

Everything LCM-internal lives in domain/. The boundary between
"things HA discovers/calls" and "things LCM owns" is now visible
from the filesystem.

Relative-import shifts:
* Within moved files: .const -> ..const, .domain.X -> .X, .models
  -> .models (sibling), .providers -> ..providers (TYPE_CHECKING)
* Within existing domain/ files: ..models -> .models (sibling),
  ..util -> .util (sibling) where applicable

External importers across providers/, root files, and tests
updated; string-based test mocks
(custom_components.lock_code_manager.models.X etc.) retargeted.
