+++ custom_components/lock_code_manager/binary_sensor.py
@@ -13,11 +13,11 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import ATTR_ACTIVE, ATTR_IN_SYNC, ATTR_SYNC_STATUS
 from .domain.coordinator import LockUsercodeUpdateCoordinator
+from .domain.models import LockCodeManagerConfigEntry
 from .domain.sync import SlotSyncManager
+from .domain.util import get_slot_coordinator
 from .entity import BaseLockCodeManagerCodeSlotPerLockEntity, BaseLockCodeManagerEntity
-from .models import LockCodeManagerConfigEntry
 from .providers import BaseLock
-from .util import get_slot_coordinator
 
 _LOGGER = logging.getLogger(__name__)
