+++ custom_components/lock_code_manager/providers/schlage.py
@@ -31,7 +31,7 @@ from ..domain.exceptions import (
     LockDisconnected,
     LockOperationFailed,
 )
-from ..models import SlotCredential
+from ..domain.models import SlotCredential
 from ._base import BaseLock
 from ._util import make_tagged_name as _make_tagged_name, parse_tag as _parse_tag
 from .const import LOGGER
