+++ custom_components/lock_code_manager/switch.py
@@ -11,9 +11,9 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from .domain.models import LockCodeManagerConfigEntry
 from .domain.slot_coordinator import PinRequiredError
 from .entity import BaseLockCodeManagerEntity
-from .models import LockCodeManagerConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
