+++ tests/test_callbacks.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from custom_components.lock_code_manager.callbacks import EntityCallbackRegistry
+from custom_components.lock_code_manager.domain.callbacks import EntityCallbackRegistry
 
 
 @pytest.fixture
