+++ tests/test_websocket.py
@@ -65,7 +65,7 @@ from custom_components.lock_code_manager.const import (
     CONF_SLOTS,
 )
 from custom_components.lock_code_manager.domain.exceptions import DuplicateCodeError
-from custom_components.lock_code_manager.models import SlotCredential
+from custom_components.lock_code_manager.domain.models import SlotCredential
 from custom_components.lock_code_manager.providers import BaseLock
 from custom_components.lock_code_manager.websocket import (
     SlotEntities,
